Added separate counter for outstanding transaction in data_obi_interface.
The counter previously used (from the lsu_respnse_filter) cannot be used in the case where a transfer happens for a region with both integrity and bufferable set. Then the outstanding counter can be manipulated by the write_buffer - thus the obi interface needs it own counter to know the state of the bus outside of the write buffer.

Signed-off-by: Oystein Knauserud <Oystein.Knauserud@silabs.com>