Implemented random replacement for HINT instruction (c.slli with rd==x0 and immediate != 0).
Replacement instruction is taken from the random instruction generation within the dummy module.
Hints are replaced by add/and/mul/bltu, but will show up on RVFI as c.slli.

Removed lfsr_lockup[0] as a cause of counter reset for the dummy module, as this could violate the dummy frequency.

Signed-off-by: Oystein Knauserud <Oystein.Knauserud@silabs.com>